:sparkles: Add day 1 - part 2 solution

+++ 2016/tools/tools.go
@@ -20,7 +20,7 @@ func Split(s string) []string {
 	return fs
 }
 
-func Abs(x int) int {
+func Abs(x float64) float64 {
 	if x < 0 {
 		return -x
 	}
